refactor: move Identity based entities to the usermanagement package
